extract some functions from the detail screens
