+++ source/views/sis/student-work/detail.android.js
@@ -5,9 +5,10 @@ import {fastGetTrimmedText} from '../../../lib/html'
 import {Card} from '../../components/card'
 import moment from 'moment'
 import openUrl from '../../components/open-url'
-import Communications from 'react-native-communications'
 import * as c from '../../components/colors'
 import type {JobType} from './types'
+import {openEmail, fixupEmailFormat} from './email'
+import {parseLinks} from './links'
 
 const styles = StyleSheet.create({
   name: {
@@ -62,21 +63,22 @@ function renderTitle(title: string, category: string) {
 function renderContact(
   title: string,
   office: string,
-  contact: string,
+  contactName: string,
   email: string,
 ) {
-  const trimmedOffice = fastGetTrimmedText(office)
-  const trimmedContact =
-    fastGetTrimmedText(contact) || email
-  return trimmedOffice || trimmedContact
+  office = fastGetTrimmedText(office)
+  email = fixupEmailFormat(email)
+  contactName =
+    fastGetTrimmedText(contactName) || email
+  return (office || contactName)
     ? <Card header="Contact" style={styles.card}>
         <Text
           style={styles.cardBody}
           onPress={() => openEmail(email, title)}
         >
-          {trimmedContact} {email ? `(${fixupEmailFormat(email)})` : ''}
+          {contactName} {email ? `(${email})` : ''}
           {'\n'}
-          {trimmedOffice}
+          {office}
         </Text>
       </Card>
     : null
@@ -155,32 +157,6 @@ function renderLastUpdated(lastModified: string) {
         {moment(lastModified, 'YYYY/MM/DD').calendar()}
       </Text>
     : null
-}
-
-function fixupEmailFormat(email: string) {
-  if (!/@/.test(email)) {
-    // No @ in address ... e.g. smith
-    return `${email}@stolaf.edu`
-  } else if (/@$/.test(email)) {
-    // @ at end ... e.g. smith@
-    return `${email}stolaf.edu`
-  } else {
-    // Defined address ... e.g. smith@stolaf.edu
-    return email
-  }
-}
-
-function openEmail(email: string, subject: string) {
-  let address = fixupEmailFormat(email)
-  Communications.email([address], null, null, subject, '')
-}
-
-function parseLinks(data: string) {
-  const allLinks = data.split(' ')
-  if (!allLinks.length) {
-    return []
-  }
-  return allLinks.filter(w => /^https?:\/\//.test(w))
 }
 
 export default function JobDetailView({job}: {job: JobType}) {
